// frontend/src/app/models/user.ts
export class User {
  firstname: string;
  lastname: string;
  username: string;
  password: string;
  contact: string;
  email: string;
  type: number;
  status: string;
  org_adress: string;
  org_name: string;
  org_number: string;
  // profilna slika
  likes: Array<String>;
  studios: Array<String>;
  icon: string;
}